requirements.txt
requests
python-dotenv
websockets
cryptography
tzdata
py-clob-client
